fixes #51: Ensure mock data includes incidents
