+++ setup/setup.sh
@@ -337,12 +337,12 @@ generate_data(){
   fi
 
   if [[ ${git_system} == "1" ]]
-  then set -x; python3 ${DIR}/../data_generator/gitlab_data.py
-  set +x
+    then set -x; python3 ${DIR}/../data_generator/generate_data.py --vc_system=gitlab
+    set +x
   fi
-  if [[ ${git_system} == "2" ]]
-  then set -x; python3 ${DIR}/../data_generator/github_data.py  
-  set +x
+    if [[ ${git_system} == "2" ]]
+    then set -x; python3 ${DIR}/../data_generator/generate_data.py --vc_system=github 
+    set +x
   fi
   
 }
